Bugfix: SVG.Builder.ModularGrid was not positioning the modules of the last row correctly if it contained a number of modules different from the non-last rows.
Bugfix: Shape searching works for Shapes.StringAction.any.

Replaced ? character with * to represent "any string action". In the future, it may be worth representing this on the bottom of the fretboard with 'bottom only' indicators for shape-chart and a corresponding 'AnyStringActionMouseTrap' for shape-input.

Adjusted shape-input height to consider possible fingerless indicators at the very bottom.